mcp/requirements.txt
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
playwright>=1.40.0

